backend/requirements.txt
anthropic>=0.32.0
python-dotenv==1.0.0
Flask==3.0.0
